Taking photo & saving
The app now enables photos to be taken and instantly saves them in a
specific area.
The app also enables the user to authenticate via twitter and tweet to
their account.
These two features need to be merged but the current tweeting method
doesn't enable images to be tweeted - need to look at other options.