Theme update & TweetActivity redesign
The app theme has been updated as well as the design of TweetActivity -
this is a now a popup activity. Included in the updates is a text hint
and a clear all 'x' button for the tweet edittext. A camera icon from
Google material icons has been added to load the camera intent.
Additionally, a textview has been placed to the right of this icon that
states whether an image has been linked to the tweet. When no image is
added it displays 'No image added' but once an image has been captured,
it displays 'Image added - click to remove', and is made clickable to
remove the image from the tweet.