Toolbar update
Updated the app toolbar to feature two buttons within the overflow menu.
One button that's been implemented so far is the ability to remove the
current active twitter session from the device - unlink twitter. The
other will be implemented soon to allow the user to remove all
favourited list items.