Progress dialogs & activity configChanges
Implemented progress dialogs for uploading an image to tweet and sending
a tweet using a custom theme style. Also added
android:configChanges="orientation|screenSize" to each activity in the
manifest to ensure the same instance is retained when changing
orientation. This prevents the activity being recreated and information
being lost due to orientation changing. This has also been tested on the
fragments within MainActivity and this method allows them to retain
their instance as well.