Google Places - nearest supermarkets
Created an asynctask to get supermarkets near to a location. These are
then placed as markers on the map. This needs updating to use the
devices current location.
Some error handling has been implemented across all asynctasks on the
app, and a progress bar has been added to display when nearby places are
loading.