Updated getting places by location
Changes have been made to the way places are acquired. The user now
presses a button to get places; this firstly acquires the devices
current location, then starts the asynctask to get places near that
location. Error handling has been implemented for when the user double
presses the button, checking whether location updates are still on and
whether the asynctask is still running - displaying toasts accordingly.
Also location updates are cancelled if still running when onStop is
called, but if the places asynctask is still running, this runs  in the
background until it either downloads data or timesout.