Recipe search & display in new activity
The app now allows for interaction with an api. Interaction with the api
accepts a search field, recipe type and amount of recipes requested.
The asynctask downloads the requested data then loops through the
recipes id's downloading the corresponding thumbnail images. These are
then cached in internal storage.
The data is tied to a parcelable object constructor (so an arraylist can
be sent via an intent). Once all downloaded, the intent loads a new
activity with the arraylist in the intent. This is then loaded into a
custom arrayadapter onCreate of the new activity, populating the
listivew.