Recently viewed functionality (Using SQLite)
Whenever a recipe is opened fully, the data is saved within a SQLite
database (or if already exists, is updated). This allows the user to
press the "view recently viewed" button to view recipes they've already
viewed. This doesn't require an internet connection as they're cached on
the device and will only be deleted once the user presses the "clear
recently viewed" button.
In regards to images, when the user opens a recipe, the temporary images
are re-saved with a different file path (removing "_temp" from the end).
This prevents them being deleted, as the existing temporary images are
deleted based on whether the file name contains thumbnail_temp or
full_size_temp. Temporary images are deleted once the
SearchResultsActivity is destroyed.