SQLite methods on Asynctask
Moved all SQLite functions onto asynctasks and added progress dialogs
for each.
Added timestamp field to recipe info table and when getting data from
this time, ordering by the timestamp field in descending order. This
allows the cached list to show the most recently viewed at the top of
the list, even if the latest viewed recipe has been viewed multiple
times.
Removed the base activity as decided the options menu is only necessary
on main activity.
Added a check on whether external storage is readable and writable
before allowing the user to take a picture for twitter.